# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synveda",
...
    "license": {
      "name": "Proprietary"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This gateway"
    }
  ],
  "paths": {
...
    "/v1/sessions": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "`GET /v1/sessions` — the runs this caller may read, newest first.",
        "operationId": "list_sessions",
        "parameters": [
          {
            "name": "scope_id",
            "in": "query",
            "description": "Only runs at or under this governed scope — the subtree, so a\nworkspace's scope lists its projects' runs.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "description": "Only runs in this workspace.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Only runs in this project.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Only runs in this state: `active`, `ending`, `ended`, `abandoned` or\n`failed`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client_name",
            "in": "query",
            "description": "Only runs opened by this agent client, matched exactly — `claude-code`,\n`zed`, `mcp`. The label the client named itself with, never a prefix.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "principal_id",
            "in": "query",
            "description": "Only runs opened by this token subject.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "started_after",
            "in": "query",
            "description": "Only runs started at or after this instant (RFC 3339).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "started_before",
            "in": "query",
            "description": "Only runs started strictly before this instant (RFC 3339).\n\nHalf-open with `started_after`, so two adjacent days cover every run\nexactly once rather than sharing the one that started at midnight.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The previous page's `next_cursor`. Opaque — pass it back unchanged.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many rows to serve: 1–200, default 50.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The sessions this caller may read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionList"
                }
              }
            }
          },
          "400": {
            "description": "A filter outside its bounds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The PDP denied `session.read`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such scope in this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "`POST /v1/sessions` — open a run.",
        "operationId": "open_session",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Required. A unique value per request, reused verbatim on retry.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenSessionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "This key already opened this session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "201": {
            "description": "Opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body, or no `Idempotency-Key`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The PDP denied `session.write`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such workspace or project in this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The workspace is archived, the harness id is taken, or the key was reused for a different request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/sessions/{session_id}": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "`GET /v1/sessions/{session_id}`.",
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "The session's id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "403": {
            "description": "The PDP denied `session.read`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such session in this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The taxonomy error body, declared for the OpenAPI document.\n\nA schema-only mirror of `synveda_types::Error`'s serialised form, which is\n`{\"kind\": \"...\", ...}` with a per-variant remainder. It exists because the\ncontract has to say what a 4xx body looks like and the taxonomy lives two\ncrates down, where `utoipa` deliberately does not reach — the OpenAPI\nderive is a property of the surface, and `synveda-types` is not one.",
        "required": [
          "kind"
        ],
        "properties": {
          "action": {
            "type": [
              "string",
              "null"
            ],
            "description": "`policy_denied` only: the action that was attempted."
          },
          "entity": {
            "type": [
              "string",
              "null"
            ],
            "description": "`not_found` only: what was looked up."
          },
          "kind": {
            "type": "string",
            "description": "The stable machine-readable code — `invalid`, `conflict`, `not_found`,\n`policy_denied`, …"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Present on most variants; what went wrong, safe to show a caller."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "`policy_denied` only: which policy produced the denial."
          },
          "resource": {
            "type": [
              "string",
              "null"
            ],
            "description": "`policy_denied` only: what was acted on."
          }
        }
      },
...
      "SessionList": {
        "type": "object",
        "description": "The session listing, one page of it.",
        "required": [
          "sessions"
        ],
        "properties": {
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where the next page resumes, or absent when this is the last one\n(CPR-11, ADR-0077 decision 1).\n\nOpaque: pass it back as `cursor` and nothing else. It replaced CPR-10's\n`truncated` boolean, which could say *that* an answer was cut short and\ncould not say where to continue — so a reader who wanted the run from\nlast Tuesday had no way to reach it.\n\nA page may be **empty and still carry one**: rows are filtered by the\nPDP after they are scanned, so a page whose candidates this caller may\nnot read serves nothing and still says where to continue. That is the\nhonest shape — the alternative is a server that keeps scanning until it\nfills a page, which is unbounded work driven by somebody else's rows."
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionView"
            },
            "description": "The sessions this caller may read, newest first."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "An OIDC access token from `synveda login`, or the console session cookie the gateway sets on its own origin."
      }
    }
  },
...
}